Isolating createKeyTweaker function
