+++ packages/pls-bitcoin/index.ts
@@ -7,7 +7,6 @@ export interface UTXO {
 	value: number;
 }
 import type { ECPairInterface } from "ecpair";
-import { ECPairFactory } from "ecpair";
 import {
 	script,
 	type Network,
@@ -16,61 +15,10 @@ import {
 } from "bitcoinjs-lib";
 import { PubkeysSchema } from "pls-core";
 import { z } from "zod";
+import { createKeyTweaker } from "./createKeyTweaker.js";
 
-import * as ecc from "tiny-secp256k1";
-
-const ECPair = ECPairFactory(ecc);
-
-type CreateKeyTweakerArgs = {
-	pubkey: Buffer;
-	privkey?: Buffer;
-}
-
-export function createKeyTweaker({ pubkey, privkey }: CreateKeyTweakerArgs) {
-	return {
-		pubkey,
-		privkey,
-		tweakPubkey(tweak: Buffer) {
-			const xOnlyPubkey = toXOnly(this.pubkey);
-
-			const tweakedPubkey = ecc.xOnlyPointAddTweak(xOnlyPubkey, tweak);
-
-			if (!tweakedPubkey)
-				throw new Error('tweak fail');
-
-			const parityByte = Buffer.from([
-				tweakedPubkey.parity === 0 ? 0x02 : 0x03,
-			]);
-
-			return Buffer.concat([
-				parityByte,
-				Buffer.from(tweakedPubkey.xOnlyPubkey),
-			]);
-		},
-		tweakPrivkey(tweak: Buffer) {
-			if (!this.privkey)
-				throw new Error('private key is not present');
-
-			const pubkey = this.pubkey;
-
-			const hasOddY = pubkey[0] === 3 || (pubkey[0] === 4 && ((pubkey[64] || 0) & 1) === 1);
-
-			const privateKey = hasOddY ? ecc.privateNegate(this.privkey) : this.privkey;
-
-			const tweakedPrivkey = ecc.privateAdd(privateKey, tweak);
-
-			if (!tweakedPrivkey)
-				throw new Error('tweak fail');
-
-			return Buffer.from(tweakedPrivkey);
-		},
-		tweakEcpair(tweak: Buffer) {
-			if (this.privkey)
-				return ECPair.fromPrivateKey(this.tweakPrivkey(tweak));
-
-			return ECPair.fromPublicKey(this.tweakPubkey(tweak));
-		},
-	}
+export {
+	createKeyTweaker,
 }
 
 const TaprootV0CollateralSchema = {
